requirements.txt
anthropic>=0.40.0
ddgs>=0.1.0
python-dotenv>=1.0.0
rich>=13.0.0
requests>=2.31.0
